+++ changelog.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Minor update in genloader to better support older distros of Raspbian
 - Updated response to remote email commands to avoid email loop problem
 - Added feature request to allow external temperature sensors to be displayed as gauges
+- Update for horizontal 120 gal tanks in mopeka sensor add on
 
 ## V1.18.15 -2022-09-18
 - Correct issue with power graph